# backend/openapi.json
{"openapi":"3.0.1","info":{"title":"ABC Telecom - Postpaid Billing System API","description":"API documentation for ABC Telecom Postpaid Billing System","contact":{"name":"ABC Telecom Dev Team","email":"devops@abctelecom.example"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Endpoints for user registration and login"},{"name":"Payments","description":"Make and list payments for invoices"},{"name":"Admin Users","description":"Admin operations for user management"},{"name":"Invoices","description":"Create and view invoices"},{"name":"Services","description":"Manage telecom services"},{"name":"Usage","description":"Usage records for services"},{"name":"Customers","description":"Customer profiles and services"}],"paths":{"/api/users/{id}":{"get":{"tags":["Admin Users"],"summary":"Get user by id (admin only)","operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Admin Users"],"summary":"Update user by id (admin only)","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Admin Users"],"summary":"Delete user by id (admin only)","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/services/{id}":{"put":{"tags":["Services"],"summary":"Update a service by id","operationId":"updateService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelecomService"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Services"],"summary":"Delete a service by id","operationId":"deleteService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/services/{id}/usage":{"get":{"tags":["Usage"],"summary":"Get usage records for a service","operationId":"getUsage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsageRecord"}}}}}}},"post":{"tags":["Usage"],"summary":"Add a usage record for a service","operationId":"addUsage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsageRecord"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/invoices/{id}/payments":{"get":{"tags":["Payments"],"summary":"List payments for an invoice","operationId":"getPayments","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Payment"}}}}}}},"post":{"tags":["Payments"],"summary":"Make a payment for an invoice","operationId":"makePayment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/customers/{id}/services":{"get":{"tags":["Customers"],"summary":"List services for a customer","operationId":"getServices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TelecomService"}}}}}}},"post":{"tags":["Customers"],"summary":"Create a telecom service for a customer","operationId":"createService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelecomService"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/customers/{id}/invoices":{"get":{"tags":["Invoices"],"summary":"List invoices for a customer","operationId":"getInvoices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Invoice"}}}}}}},"post":{"tags":["Invoices"],"summary":"Generate an invoice for a customer over a period","operationId":"createInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"start","in":"query","required":true,"schema":{"type":"string"}},{"name":"end","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"},"examples":{"example":{"description":"example","value":{"username":"user_example","password":"P@ssw0rd123","email":"user+unique@example.com","role":"CUSTOMER"}}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and receive a JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"},"examples":{"example":{"description":"example","value":{"username":"user_example","password":"P@ssw0rd123"}}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/invoices/{id}":{"get":{"tags":["Invoices"],"summary":"Get an invoice by id","operationId":"getInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/customers/{id}":{"get":{"tags":["Customers"],"summary":"Get customer profile by id","operationId":"getCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"passwordHash":{"type":"string","writeOnly":true},"email":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"TelecomService":{"type":"object","properties":{"serviceId":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"serviceType":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UsageRecord":{"type":"object","properties":{"usageId":{"type":"integer","format":"int64"},"serviceId":{"type":"integer","format":"int64"},"usageDate":{"type":"string","format":"date-time"},"usageAmount":{"type":"number","format":"double"},"unit":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Payment":{"type":"object","properties":{"paymentId":{"type":"integer","format":"int64"},"invoiceId":{"type":"integer","format":"int64"},"paymentDate":{"type":"string","format":"date-time"},"amount":{"type":"number","format":"double"},"paymentMethod":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Invoice":{"type":"object","properties":{"invoiceId":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"billingPeriodStart":{"type":"string","format":"date-time"},"billingPeriodEnd":{"type":"string","format":"date-time"},"totalAmount":{"type":"number","format":"double"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date-time"}}}}}}